Refactor common range styles into StyledRange
